+++ x/token/event.go
@@ -2,24 +2,50 @@ package token
 
 import (
 	"fmt"
+	"strings"
 
 	sdk "github.com/line/lbm-sdk/types"
 )
 
-func NewEventIssueToken(e EventIssue, grantee, to sdk.AccAddress, amount sdk.Int) sdk.Event {
-	eventType := EventType_IssueToken.String()
-	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_Name:       e.Name,
-		AttributeKey_Symbol:     e.Symbol,
-		AttributeKey_ImageURI:   e.Uri,
-		AttributeKey_Meta:       e.Meta,
-		AttributeKey_Decimals:   fmt.Sprintf("%d", e.Decimals),
-		AttributeKey_Mintable:   fmt.Sprintf("%t", e.Mintable),
+const (
+	prefixEventType    = "EVENT_TYPE_"
+	prefixAttributeKey = "ATTRIBUTE_KEY_"
+)
 
-		AttributeKey_Owner:  grantee.String(),
-		AttributeKey_To:     to.String(),
-		AttributeKey_Amount: amount.String(),
+func (x EventType) String() string {
+	lenPrefix := len(prefixEventType)
+	return strings.ToLower(EventType_name[int32(x)][lenPrefix:])
+}
+
+func EventTypeFromString(name string) EventType {
+	eventTypeName := prefixEventType + strings.ToUpper(name)
+	return EventType(EventType_value[eventTypeName])
+}
+
+func (x AttributeKey) String() string {
+	lenPrefix := len(prefixAttributeKey)
+	return strings.ToLower(AttributeKey_name[int32(x)][lenPrefix:])
+}
+
+func AttributeKeyFromString(name string) AttributeKey {
+	attributeKeyName := prefixAttributeKey + strings.ToUpper(name)
+	return AttributeKey(AttributeKey_value[attributeKeyName])
+}
+
+func NewEventIssueToken(e EventIssue, grantee, to sdk.AccAddress, amount sdk.Int) sdk.Event {
+	eventType := EventTypeIssueToken.String()
+	attributes := map[AttributeKey]string{
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyName:       e.Name,
+		AttributeKeySymbol:     e.Symbol,
+		AttributeKeyImageURI:   e.Uri,
+		AttributeKeyMeta:       e.Meta,
+		AttributeKeyDecimals:   fmt.Sprintf("%d", e.Decimals),
+		AttributeKeyMintable:   fmt.Sprintf("%t", e.Mintable),
+
+		AttributeKeyOwner:  grantee.String(),
+		AttributeKeyTo:     to.String(),
+		AttributeKeyAmount: amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -31,12 +57,12 @@ func NewEventIssueToken(e EventIssue, grantee, to sdk.AccAddress, amount sdk.Int
 }
 
 func NewEventMintToken(e EventMinted) sdk.Event {
-	eventType := EventType_MintToken.String()
+	eventType := EventTypeMintToken.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_From:       e.Operator,
-		AttributeKey_To:         e.To,
-		AttributeKey_Amount:     e.Amount.String(),
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyFrom:       e.Operator,
+		AttributeKeyTo:         e.To,
+		AttributeKeyAmount:     e.Amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -48,11 +74,11 @@ func NewEventMintToken(e EventMinted) sdk.Event {
 }
 
 func NewEventBurnToken(e EventBurned) sdk.Event {
-	eventType := EventType_BurnToken.String()
+	eventType := EventTypeBurnToken.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_From:       e.From,
-		AttributeKey_Amount:     e.Amount.String(),
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyFrom:       e.From,
+		AttributeKeyAmount:     e.Amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -64,12 +90,12 @@ func NewEventBurnToken(e EventBurned) sdk.Event {
 }
 
 func NewEventBurnTokenFrom(e EventBurned) sdk.Event {
-	eventType := EventType_BurnTokenFrom.String()
+	eventType := EventTypeBurnTokenFrom.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_Proxy:      e.Operator,
-		AttributeKey_From:       e.From,
-		AttributeKey_Amount:     e.Amount.String(),
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyProxy:      e.Operator,
+		AttributeKeyFrom:       e.From,
+		AttributeKeyAmount:     e.Amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -81,10 +107,10 @@ func NewEventBurnTokenFrom(e EventBurned) sdk.Event {
 }
 
 func NewEventModifyToken(e EventModified) []sdk.Event {
-	eventType := EventType_ModifyToken.String()
+	eventType := EventTypeModifyToken.String()
 	new := []sdk.Event{
 		sdk.NewEvent(eventType,
-			sdk.NewAttribute(AttributeKey_ContractID.String(), e.ContractId),
+			sdk.NewAttribute(AttributeKeyContractID.String(), e.ContractId),
 		),
 	}
 
@@ -97,12 +123,12 @@ func NewEventModifyToken(e EventModified) []sdk.Event {
 }
 
 func NewEventTransfer(e EventSent) sdk.Event {
-	eventType := EventType_Transfer.String()
+	eventType := EventTypeTransfer.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_From:       e.From,
-		AttributeKey_To:         e.To,
-		AttributeKey_Amount:     e.Amount.String(),
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyFrom:       e.From,
+		AttributeKeyTo:         e.To,
+		AttributeKeyAmount:     e.Amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -114,13 +140,13 @@ func NewEventTransfer(e EventSent) sdk.Event {
 }
 
 func NewEventTransferFrom(e EventSent) sdk.Event {
-	eventType := EventType_TransferFrom.String()
+	eventType := EventTypeTransferFrom.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_Proxy:      e.Operator,
-		AttributeKey_From:       e.From,
-		AttributeKey_To:         e.To,
-		AttributeKey_Amount:     e.Amount.String(),
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyProxy:      e.Operator,
+		AttributeKeyFrom:       e.From,
+		AttributeKeyTo:         e.To,
+		AttributeKeyAmount:     e.Amount.String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -132,14 +158,14 @@ func NewEventTransferFrom(e EventSent) sdk.Event {
 }
 
 func NewEventGrantPermToken(e EventGrant) sdk.Event {
-	eventType := EventType_GrantPermToken.String()
+	eventType := EventTypeGrantPermToken.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_To:         e.Grantee,
-		AttributeKey_Perm:       e.Permission,
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyTo:         e.Grantee,
+		AttributeKeyPerm:       LegacyPermission(e.Permission).String(),
 	}
 	if e.Granter != e.Grantee {
-		attributes[AttributeKey_From] = e.Granter
+		attributes[AttributeKeyFrom] = e.Granter
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -151,11 +177,11 @@ func NewEventGrantPermToken(e EventGrant) sdk.Event {
 }
 
 func NewEventRevokePermToken(e EventAbandon) sdk.Event {
-	eventType := EventType_RevokePermToken.String()
+	eventType := EventTypeRevokePermToken.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_From:       e.Grantee,
-		AttributeKey_Perm:       e.Permission,
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyFrom:       e.Grantee,
+		AttributeKeyPerm:       LegacyPermission(e.Permission).String(),
 	}
 
 	new := sdk.NewEvent(eventType)
@@ -167,11 +193,11 @@ func NewEventRevokePermToken(e EventAbandon) sdk.Event {
 }
 
 func NewEventApproveToken(e EventAuthorizedOperator) sdk.Event {
-	eventType := EventType_ApproveToken.String()
+	eventType := EventTypeApproveToken.String()
 	attributes := map[AttributeKey]string{
-		AttributeKey_ContractID: e.ContractId,
-		AttributeKey_Approver:   e.Holder,
-		AttributeKey_Proxy:      e.Operator,
+		AttributeKeyContractID: e.ContractId,
+		AttributeKeyApprover:   e.Holder,
+		AttributeKeyProxy:      e.Operator,
 	}
 
 	new := sdk.NewEvent(eventType)